feat(firebase): Setup Firebase Functions with TypeScript
- Created setup script for Firebase Functions environment
- Migrated to TypeScript for better type safety and maintainability
- Added comprehensive setup documentation
- Configured ESLint and Prettier for code quality
- Updated Firebase configuration for TypeScript support
