+++ functions/.eslintrc.js
@@ -1,0 +1,15 @@
+module.exports = {
+  root: true,
+  env: {
+    es6: true,
+    node: true,
+  },
+  extends: [
+    "eslint:recommended",
+    "google",
+  ],
+  rules: {
+    quotes: ["error", "double"],
+    "max-len": ["error", { "code": 100 }],
+  },
+};
